+++ src/chanjo2/endpoints/overview.py
@@ -1,12 +1,14 @@
 from os import path
 
-from fastapi import APIRouter, Request
+from fastapi import APIRouter, Request, Depends
 from fastapi.responses import HTMLResponse
 from fastapi.templating import Jinja2Templates
+from sqlmodel import Session
 
-from chanjo2.demo import DEMO_OVERVIEW_QUERY_DATA
-from chanjo2.meta.handle_overview_content import get_overview_data
-from chanjo2.models.pydantic_models import GeneralReportQuery
+from chanjo2.dbutil import get_session
+from chanjo2.demo import DEMO_COVERAGE_QUERY_DATA
+from chanjo2.meta.handle_report_contents import get_report_data
+from chanjo2.models.pydantic_models import ReportQuery
 
 
 def get_templates_path() -> str:
@@ -20,15 +22,19 @@ router = APIRouter()
 
 
 @router.get("/overview/demo", response_class=HTMLResponse)
-async def demo_overview(request: Request):
+async def demo_overview(request: Request, db: Session = Depends(get_session)):
     """Return a demo genes overview page over a list of genes for a list of samples."""
 
-    overview_query = GeneralReportQuery(**DEMO_OVERVIEW_QUERY_DATA)
-    overview_content: dict = get_overview_data(query=overview_query)
+    overview_query = ReportQuery(**DEMO_COVERAGE_QUERY_DATA)
+    overview_content: dict = get_report_data(
+        query=overview_query, session=db, is_overview_report=True
+    )
     return templates.TemplateResponse(
         "overview.html",
         {
             "request": request,
             "extras": overview_content["extras"],
+            "levels": overview_content["levels"],
+            "incomplete_coverage_rows": overview_content["incomplete_coverage_rows"],
         },
     )