+++ src/chanjo2/endpoints/report.py
@@ -29,7 +29,9 @@ async def demo_report(request: Request, db: Session = Depends(get_session)):
     """Return a demo coverage report over a list of genes for a list of samples."""
 
     report_query = ReportQuery(**DEMO_COVERAGE_QUERY_DATA)
-    report_content: Dict = get_report_data(query=report_query, session=db)
+    report_content: Dict = get_report_data(
+        query=report_query, session=db, is_overview_report=False
+    )
     return templates.TemplateResponse(
         "report.html",
         {
@@ -52,7 +54,9 @@ async def report(
     request: Request, report_query: ReportQuery, db: Session = Depends(get_session)
 ):
     """Return a coverage report over a list of genes for a list of samples."""
-    report_content: Dict = get_report_data(query=report_query, session=db)
+    report_content: Dict = get_report_data(
+        query=report_query, session=db, is_overview_report=False
+    )
     return templates.TemplateResponse(
         "report.html",
         {