+++ tests/src/chanjo2/meta/test_handle_report_contents.py
@@ -10,6 +10,7 @@ from chanjo2.meta.handle_report_contents import (
     get_report_level_completeness_rows,
     get_report_completeness_rows,
     get_missing_genes_from_db,
+    get_genes_overview_incomplete_coverage_rows,
 )
 from chanjo2.models.sql_models import Gene as SQLGene
 from chanjo2.models.sql_models import Transcript as SQLTranscript
@@ -102,3 +103,40 @@ def test_get_missing_genes_from_db(
     # THEN the get_missing_genes_from_db function should return the expected error, containing description and missing IDs
     assert missing_gene_error[0] == "Gene IDs not found in the database"
     assert missing_gene_error[1]
+
+
+def test_get_genes_overview_incomplete_coverage_rows(
+    demo_session: sessionmaker,
+    samples_d4_list: List[Tuple[str, D4File]],
+    demo_genes_37: List[SQLGene],
+):
+    """Test the function that returnes the lines of the genes coverage overview page."""
+
+    # GIVEN coverage stats for a list of cases at the transcript level
+    samples_coverage_stats: Dict[str, List[GeneCoverage]] = {
+        sample: get_sample_interval_coverage(
+            db=demo_session,
+            d4_file=d4_file,
+            genes=demo_genes_37,
+            interval_type=SQLTranscript,
+            completeness_thresholds=DEFAULT_COMPLETENESS_LEVELS,
+        )
+        for sample, d4_file in samples_d4_list
+    }
+    assert samples_coverage_stats
+
+    # THEN it should be possible to retrieve stats for the transcripts that are not fully covered at the given threshold
+    genes_overview_lines: List[
+        Tuple[Union[str, int, float]]
+    ] = get_genes_overview_incomplete_coverage_rows(
+        samples_coverage_stats=samples_coverage_stats,
+        interval_type=SQLTranscript,
+        cov_level=DEFAULT_COMPLETENESS_LEVELS[0],
+    )
+
+    # THEN each transcript line should contain the expected values
+    for gene, transcript_id, sample, completeness in genes_overview_lines:
+        assert isinstance(gene, str)
+        assert isinstance(transcript_id, str)
+        assert sample == samples_d4_list[0][0]
+        assert completeness < 1